Actualizacion V 0.4

+++ database/migrations/2021_08_10_214240_create_placas_table.php
@@ -22,7 +22,7 @@ class CreatePlacasTable extends Migration
             $table->date('baja');
             $table->date('fecha_pago');
             $table->date('fecha_estimada');
-            $table->integer('monto');
+            $table->decimal('monto');
             $table->integer('año');
             $table->string('cambio_propietario');
             $table->date("fecha_cambio_p")->nullable();
