+++ database/migrations/2021_08_11_184037_create_polizas_table.php
@@ -22,7 +22,7 @@ class CreatePolizasTable extends Migration
             $table->date('vencimiento');
             $table->date('fecha_pago');
             $table->date('fecha_estimada');
-            $table->integer('monto');
+            $table->decimal('monto');
             $table->integer('año');
             $table->string('endoso');
             $table->string('concepto_endoso')->nullable();
