+++ database/migrations/2021_08_16_150425_create_verificacion_fs_table.php
@@ -19,7 +19,7 @@ class CreateVerificacionFsTable extends Migration
             $table->integer('verificacion');
             $table->date('fecha_pago');
             $table->date('fecha_estimada');
-            $table->integer('monto');
+            $table->decimal('monto');
             $table->string('estatus');
             $table->string('archivo');
             $table->unsignedBigInteger('id_vehiculo');
