+++ database/migrations/2021_08_25_163026_create_tarjetacs_table.php
@@ -23,7 +23,7 @@ class CreateTarjetacsTable extends Migration
             $table->date('vencimiento');
             $table->date('fecha_pago');
             $table->date('fecha_estimada');
-            $table->integer('monto');
+            $table->decimal('monto');
             $table->integer('año');
             $table->unsignedBigInteger('id_vehiculo');
             $table->foreign('id_vehiculo')->references('id')->on('vehiculos');
